+++ src/Commands/MakeModel.php
@@ -11,7 +11,7 @@ class MakeModel extends DomainGeneratorCommand
      *
      * @var string
      */
-    protected $signature = 'ddd:make:model {domain} {name}';
+    protected $signature = 'ddd:model {domain} {name}';
 
     /**
      * The console command description.
@@ -39,19 +39,14 @@ class MakeModel extends DomainGeneratorCommand
         $parts = str($baseModel)->explode('\\');
         $baseModelName = $parts->last();
         $baseModelPath = $this->getPath($baseModel);
-        // dd($baseModelPath);
 
         if (! file_exists($baseModelPath)) {
             $this->warn("Base model {$baseModel} doesn't exist, generating...");
-
-            // dd($baseModel, $baseModelName);
 
             $this->call(MakeBaseModel::class, [
                 'domain' => 'Shared',
                 'name' => $baseModelName,
             ]);
-        } else {
-            // dd('file exists', $baseModelPath);
         }
 
         parent::handle();