+++ src/Commands/MakeViewModel.php
@@ -11,7 +11,7 @@ class MakeViewModel extends DomainGeneratorCommand
      *
      * @var string
      */
-    protected $signature = 'ddd:make:view-model {domain} {name}';
+    protected $signature = 'ddd:view-model {domain} {name}';
 
     /**
      * The console command description.
@@ -30,5 +30,25 @@ class MakeViewModel extends DomainGeneratorCommand
     protected function getStub()
     {
         return $this->resolveStubPath('view-model.php.stub');
+    }
+
+    public function handle()
+    {
+        $baseViewModel = config('ddd.base_view_model');
+
+        $parts = str($baseViewModel)->explode('\\');
+        $baseName = $parts->last();
+        $basePath = $this->getPath($baseViewModel);
+
+        if (! file_exists($basePath)) {
+            $this->warn("Base view model {$baseViewModel} doesn't exist, generating...");
+
+            $this->call(MakeBaseViewModel::class, [
+                'domain' => 'Shared',
+                'name' => $baseName,
+            ]);
+        }
+
+        parent::handle();
     }
 }